Allow configuration of popup size and position
This is a quick-and-dirty implementation of allowing the user to
configure the popup size and position.

Popup size is set using `@extrakto_popup_size`, which takes a
comma-separated argument specifying width and height (in that order).

Popup position is set similarly using `@extrakto_popup_position` for the
x and y positions of the popup window, as in the description of
`display-menu` and `display-popup` in `man tmux`.

This implementation lacks (and sorely needs) input validation. It will
break as soon as the user chooses to provide only one of the width
or height. The same goes for the x/y positions.

I'm not a huge fan of the way I use `awk` to parse the options (do I
really need an `echo` there?), but it should be portable across
different versions of awk.

I'll update this pull request when I find time to work out the kinks I
outlined above.