feat: add @extrakto_alt and @extrakto_prefix_name options (#133)
* feat: add @extrakto_alt and @extrakto_prefix_name options

By default, extrakto_plugin is only showing alternative options to
filters and prefixing the results with the name filter whenever we
trigger the `all` filters option.

extrakto provides us the option to enable or disable both the filter
alternatives and the prefixing of the names.

This commits allows this option our plugin

* refactor: apply latest black formatting

* fix: take into account prefix_name when selecting entries

* feat: implement extrakto_alt and extrakto_prefix_name logic using all, any and none values